+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ```diff
 + Added a LockSize option `Tab:UpdateWarningBox()` to set the maximum size of the warning box to 3.25 size of the Tab Container (optional)
++ Added support for mouse button 3 (middle click)
 ```
 
 ### 17.07.2025